remove emacs ohmyzsh plugin and create my own alias
The plugin was causing an issue because it created an alias for
emacsclient that started a daemon in a way that isn't compatible with how I have
things set up. I realized I wasn't using any other feature of the plugin
so I just removed it and made my own alias for emacsclient.